Don't remove file if not there

+++ dcjulian29.groups.personal/tools/chocolateyInstall.ps1
@@ -145,4 +145,6 @@ if (-not (Test-Path -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt")) 
   netsh int ipv4 show excludedportrange udp
 }
 
-Remove-Item -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt" -Force
+if (Test-Path -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt") {
+  Remove-Item -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt" -Force
+}
