+++ dcjulian29.groups.personal/tools/chocolateyUninstall.ps1
@@ -1,1 +1,3 @@
-Remove-Item -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt" -Force
+if (Test-Path -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt") {
+  Remove-Item -Path "${env:TEMP}\dcjulian29.groups.personal.update.txt" -Force
+}
